Updated the "q_balance" equation to account for possible ramp scheduling.
NOTE! Since the state terms use a different time indexing in the non-ramp-schedule formulation than the rest of the terms, the ramp scheduling also has to refer to different time indeces. However as far as I know the indexing in the ramp-schedule case makes the indexing identical between state and other terms again.