Separated the absolute reserve-use concious state variable constraints from the slack variable constraints.
The revised stave variable reserve constraints feature improved accounting for the dynamics of the systems in the form of the self-discharge and diffusion coefficients being accounted for. The state slack variable constraints are currently rather basic, and do not feature the use of reserves in any way, resulting in the reserve provisions not being penalized for possibly exceeding desired state limits.