Intermediate commit for implementing time series form unit parameters. HAS NOT BEEN TESTED WITH ACTUAL TIME SERIES DATA  AS OF YET!
First steps for implementing the possibility of time series form input data for units, required for e.g. heat pumps. The idea is to use a flag "useTimeseries" in p_unit to easily disable the use of time series when necessary. However, the flag is not parameter dependent, meaning that if the flag is set all available time series data for each unit will be used. The code should use time series data only when able, and supplement missing values with constant parameter data, which should make it possible (if not advised) to feed incomplete time series as input.