Fixed a small issue, where there would be divisions by zero if the optimizaiton horizon was set to an end of a defined interval, which wasn't the last defined interval.
This is nothing crucial, as intervals past the optimization horizon probably shouldn't be defined at all. However, makes it a little more convenient to experiment with the model settings. As the last step is defined separately in order to correctly set the end point state boundaries anyway, there is no need to account for it in the interval loop.