Added the final state and transfer variables to the results, as well as a cautionary reminder regarding the way the length of the time series form data is determined in the model. This issue will need to be addressed at some point.
The length of the time series data is currently determined as the maximum length of any provided "ts_energyDemand" input. However, this time series is not required for all modelling purposes, as it can be substituted by the "ts_absolute" in most (if not all) cases. If "ts_energyDemand" is not included in the input, the set "ct(t)" will not function properly, causing (at the very least) the looping of the timeseries data to fail. Then again, I'm fairly certain the loop also fails if the time series don't have compatible lengths, e.g. some time series input is shorter than the maximum length of "ts_energyDemand".