Fixed an issue where the "vq_resDemand" dummy variables were generated for nodes without reserve requirements. Also tweaked the penalty terms to improve solution viability.
CAUTION! The reserve penalty seems to cause some issues, if it is of the same order of magnitude as the energy penalty. This is probably due to the multiple categories of reserves accumulating penalties faster than the energy balances at the nodes. This could result in the model purposefully violating the energy balance, if reserve requirements would need to be compromised, which is not intended behavior.