Created 3 new variables and 2 new equations to permit imposing ramp constraints and costs on units. Also generalized the "resdirection" set as an "up_down" set, also used for ramp related variables.
IMPORTANT! The dummy data in "3a_modelsInit.gms" regarding the reserve requirements needs to be changed in order for these changes to function!

The new "v_genRamp" records the change in the "v_gen" over an interval, as required in the "q_genRamp" equality constraint. The "v_genRamp" variable is generated for units with ramp restrictions, and is used for enforcing those in the model. However, in order to apply larger costs to steeper ramps, another variable called "v_genRampChange" is required, that essentially tracks the changes in the ramp speed of the units according to the "q_genRampChange" equality constraint. Since this variable is used to apply the costs, it requires the "up_down" dimension to keep track of positive and negative changes separately.

In order to make the ramping constraints as general and specific as possible, they can be imposed for each unit output separately. The old ramp related parameters were thus moved from "param_unit" under "param_gnu".