Making it possible for "maxGen" to be negative, which is a special case currently required by cooling equipment.
Cooling equipment are quite irregular when it comes to Backbone's node-unit-structure, as they consume energy in order to get rid of energy elsewhere. In this commit, the desired behavior is achieved by defining both the "maxGen" and "slope" of the cooling units as negative, which results in the input "v_gen" still being drawn as normal, but the output "v_gen" is converted into "negative energy" with the negative "slope".

Another possible approach would have been to define cooling equipment as sort of heat pumps (which they sort of are in reality), defining both the electrican grid and the cooled heat node as input nodes and constraining their input ratios. However, this would have required creating a heat sink node where the heat is transported, and giving that node the "v_spill" variable. All in all, this method was simpler for the time being.