Removed "p_uft_online_last" from use, as it is not currently working properly and possibly confuses the objective function.
At the moment, online variables are generated only for the time steps when they actually affect the v_gen variables. Thus, having this sort of incentive term to keep the units online at the end of the simulation is perhaps not necessary. The horizon extends over the t_end anyways, and the units keep on working as if they still had to keep on keeping on.