Fixing an issue where the "v_reserve" variable for input units was being counted twice in the "q_resDemand" equation. Also limited the averaging of the "ts_reserveDemand_" time series to "t_reserveLength", as it is unnecessary to average the time series over the reserve horizon.
I'm unsure why the input units were handled separately in the "q_resDemand" equation, but it seems to me that this is a relic from versions past. No (node, unit) should be able to influence the properties of other nodes directly, as these types of interactions are handled either by the "v_transfer" or conversion units, which have separate (grid, node, unit) entities on each of the nodes concerned. These are included in the (node, unit, f, t) of their respective nodes, so there should be no need to include other reserve variables to the nodes.