Added a reminder to "q_startupType" that the constraint isn't working as intended and requires changes.
The current formulation of the startup type constraint allows for multiple sub-units to hot/warm start on subsequent timesteps for each shutdown within the hot/warm startup time window. Unfortunately this issue cannot be solved simply by considering the startups within the time window as well, as this could result in forcible shutdowns when only recent startups are seen by the time window. This could be done if the entire simulated history (t_full) would be included in the equations, but this results in severe performance issues.

The current formulation still works for units with "unitCount" = 1, and since cold starts are not affected they still work as intended.