Removing an unnecessary conditional from the "q_startshut" equation, as well as a slight cosmetic change to "q_resDemand".
The condition ${ ord(t) + dt(t) > mSettings(m, 't_start') } is unnecessary, as the model can currently freely decide which power plants are online on t00 without any costs in the objective function. Thus, it should no longer be necessary to neglect this equation on the first active time step.